fix signup route

+++ api/utils/object_storage.py
@@ -3,7 +3,10 @@ from datetime import datetime
 from pathlib import Path, PurePath
 import swiftclient
 
-from utils.config import IS_DEV, IS_STAGING
+from utils.config import IS_DEV
+
+def get_storage_base_url():
+    return os.environ.get('OBJECT_STORAGE_URL')
 
 
 def swift_con():
@@ -27,55 +30,50 @@ STORAGE_DIR = Path(os.path.dirname(os.path.realpath(__file__)))\
               / '..' / 'static' / 'object_store_data'
 
 
-def local_dir(bucket, id):
-    if '/' in id:
-        idFolders = PurePath(id).parent
+def local_dir(bucket, entity_id):
+    if '/' in entity_id:
+        id_folders = PurePath(entity_id).parent
     else:
-        idFolders = ''
-    return STORAGE_DIR / bucket / idFolders
+        id_folders = ''
+    return STORAGE_DIR / bucket / id_folders
 
 
-def local_path(bucket, id):
-    return local_dir(bucket, id) / PurePath(id).name
+def local_path(bucket, entity_id):
+    return local_dir(bucket, entity_id) / PurePath(entity_id).name
 
 
-def store_public_object(bucket, id, blob, content_type, symlink_path=None):
-
+def store_public_object(bucket, entity_id, blob, content_type, symlink_path=None):
     if IS_DEV:
-        os.makedirs(local_dir(bucket, id), exist_ok=True)
+        os.makedirs(local_dir(bucket, entity_id), exist_ok=True)
 
-        file_local_path = local_path(bucket, id)
+        file_local_path = local_path(bucket, entity_id)
 
-        newTypeFile = open(str(file_local_path)+".type", "w")
-        newTypeFile.write(content_type)
+        new_type_file = open(str(file_local_path) + ".type", "w")
+        new_type_file.write(content_type)
 
-        if symlink_path and not os.path.isfile(file_local_path):
+        if symlink_path and not os.path.isfile(file_local_path) and not os.path.islink(file_local_path):
             os.symlink(symlink_path, file_local_path)
             return
 
-        newFile = open(file_local_path, "wb")
-        newFile.write(blob)
-        return
-
+        new_file = open(file_local_path, "wb")
+        new_file.write(blob)
     else:
-        # PRODUCTION OVH LIKE
         container_name = os.environ.get('OVH_BUCKET_NAME')
-        # we want to store data with a special path
-        storage_path = 'thumbs/' + id
+        storage_path = 'thumbs/' + entity_id
         swift_con().put_object(container_name,
                                storage_path,
                                contents=blob,
                                content_type=content_type)
 
 
-def delete_public_object(bucket, id):
-    lpath = local_path(bucket, id)
+def delete_public_object(bucket, entity_id):
+    lpath = local_path(bucket, entity_id)
     if os.path.isfile(lpath):
         os.remove(lpath)
 
 
-def get_public_object_date(bucket, id):
-    lpath = local_path(bucket, id)
+def get_public_object_date(bucket, entity_id):
+    lpath = local_path(bucket, entity_id)
     if not os.path.isfile(lpath):
         return None
     return datetime.fromtimestamp(os.path.getmtime(lpath))
